webapp_v2/functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.5.0
Pillow>=10.0.0